requirements.txt
# Web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Speech processing
deepgram-sdk>=3.0.0
silero-vad>=4.0.0
torch>=2.0.0  # Required for Silero VAD

# Text-to-speech
cartesia>=0.1.0  # Or use HTTP client if SDK not available

# LLM integration
httpx>=0.25.0  # For OpenRouter API calls
openai>=1.0.0  # Alternative if using OpenAI SDK

# Data validation
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Logging and observability
structlog>=23.0.0  # Optional: structured logging

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality (development)
black>=23.0.0
mypy>=1.5.0
ruff>=0.1.0

# Resume parsing
pdfplumber>=0.10.0
python-docx>=1.1.0
aiofiles>=23.0.0

# PDF generation for transcript download
reportlab>=4.0.0

